Commit de sauvegarde...

+++ branches/refactoring4/src/main/java/abstractions/evaluator/EvaluatorInterface.java
@@ -38,7 +38,7 @@ public interface EvaluatorInterface {
      * 
      * @return evaluated moves
      */
-    TreeMap<Integer, List<MutationInterface>> applyEvaluation(final List<MutationInterface> mutations);
+    TreeMap<Double, List<MutationInterface>> applyEvaluation(final List<MutationInterface> mutations);
 
     //cf TODO
     void injectContext(final ContextInterface context);
